separate ptarmd.c and main()
